+++ osutil/constrain_windows.go
@@ -4,6 +4,6 @@ func Constrain(userName, groupName, chrootDir string) error {
 	return nil
 }
 
-func ConstraintReport() string {
-	return "uid=windows gid=windows cwd=?"
+func ConstraintReport(chroot string) string {
+	return "uid=windows gid=windows cwd=" + chroot
 }